PRSD-364: Move email template models
