+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/viewModels/EmailTemplateModel.kt
@@ -1,4 +1,4 @@
-package uk.gov.communities.prsdb.webapp.viewmodel
+package uk.gov.communities.prsdb.webapp.models.viewModels
 
 interface EmailTemplateModel {
     val templateId: EmailTemplateId
